Design modifications + method
implemented tie detection.
tie detection should be called after every round,
method implemented and tested working fine,
design of the program was modified,
created a class container of all the methodes, main program should be the upper layer that handle the "Gui and I/O" all the logic of the game should be in game class.